Adjust messages_path to support having the locale files in /usr/share/locale
Linux distribution normally centralize their locale files in /usr/share/locale
as a central place.
With this commit we check if the expected locale directory is present if so we
use it, otherwise we use /usr/share/locale.

This will allow packager of wtforms in Linux distribution to place the locale
files in the expected location in their distro