Adds a validator check when creating a field
When doing *Field(*args, validators=[DummyValidator]), there is
no check on the callable received (DummyValidator). For builtin
validators, if this is not an instance, you get an incomprehensible
error from _run_validation_chain.
This commit adds a check at field's initialization time to ensure
either a callable instance object or a function has been given as
a validator by raising a meaningful error message.

Signed-off-by: mdew <adrienoyono@gmail.com>