Fix for BooleanField always evaluating to True.
When used in conjunction with Flask-WTF, `Field.process()` will be
passed a formdata object (MultiDict), with a bool value False, even when
used with `Form(obj=obj)`, which should support a raw dict.
Because this will then trigger `Field.process_formdata([False])`,
`BooleanField.process_formdata()` will evaluate to `True`, which is,
err, false. This PR correctly recognizes that case.