Use iter_choices when validating
This give a consistent output of 3 values where using just choices leads to ValueError when choices is a list of strings. By using match from iter_choices we also keep from doing the check twice so there should be no costs here except for the zip() when using lists of strings.