Added missing Spanish translations
I have added missing Spanish translations, and added missing periods for consistency.
For the translation in line 141: ```Invalid choice(s): one or more data inputs could not be coerced.```
It seems like the previous translator was unsure whether or not to use "coaccionadas" or "ajustadas". 
As "coaccionadas" is a valid programming term (but sounds awkward in everyday use), I made the decision to translate it for programmers.
Unless you disagree with my translations, I believe it should be good to merge.
Let me know if you have any questions on the changes.
Cheers!