URL validator: accept user/password, use urllib.urlparse
I was hitting an issue in a build tool that was not letting me specify
a URL to clone a git tree with a personal access token (e.g. [1]) in a
wtform URL field.

I started looking at expanding the original regex, but there are
tricks like multiple "@"'s in passwords that are hard to get right.  I
think that for this purpose, urllib.urlparse (urlparse/urlsplit
doesn't seem to matter here) will just "do the right thing".

The test-cases are expanded with some coverage of username/passwords.

[1] https://docs.gitlab.com/ee/user/profile/personal_access_tokens.html#clone-repository-using-personal-access-token